requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
Pillow==10.1.0
aiofiles==23.2.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
